+++ crates/common/common_enums/src/enums.rs
@@ -2048,262 +2048,6 @@ pub enum Country {
     Zimbabwe,
 }
 
-impl Country {
-    pub const fn from_alpha2(code: CountryAlpha2) -> Self {
-        match code {
-            CountryAlpha2::AF => Self::Afghanistan,
-            CountryAlpha2::AX => Self::AlandIslands,
-            CountryAlpha2::AL => Self::Albania,
-            CountryAlpha2::DZ => Self::Algeria,
-            CountryAlpha2::AS => Self::AmericanSamoa,
-            CountryAlpha2::AD => Self::Andorra,
-            CountryAlpha2::AO => Self::Angola,
-            CountryAlpha2::AI => Self::Anguilla,
-            CountryAlpha2::AQ => Self::Antarctica,
-            CountryAlpha2::AG => Self::AntiguaAndBarbuda,
-            CountryAlpha2::AR => Self::Argentina,
-            CountryAlpha2::AM => Self::Armenia,
-            CountryAlpha2::AW => Self::Aruba,
-            CountryAlpha2::AU => Self::Australia,
-            CountryAlpha2::AT => Self::Austria,
-            CountryAlpha2::AZ => Self::Azerbaijan,
-            CountryAlpha2::BS => Self::Bahamas,
-            CountryAlpha2::BH => Self::Bahrain,
-            CountryAlpha2::BD => Self::Bangladesh,
-            CountryAlpha2::BB => Self::Barbados,
-            CountryAlpha2::BY => Self::Belarus,
-            CountryAlpha2::BE => Self::Belgium,
-            CountryAlpha2::BZ => Self::Belize,
-            CountryAlpha2::BJ => Self::Benin,
-            CountryAlpha2::BM => Self::Bermuda,
-            CountryAlpha2::BT => Self::Bhutan,
-            CountryAlpha2::BO => Self::BoliviaPlurinationalState,
-            CountryAlpha2::BQ => Self::BonaireSintEustatiusAndSaba,
-            CountryAlpha2::BA => Self::BosniaAndHerzegovina,
-            CountryAlpha2::BW => Self::Botswana,
-            CountryAlpha2::BV => Self::BouvetIsland,
-            CountryAlpha2::BR => Self::Brazil,
-            CountryAlpha2::IO => Self::BritishIndianOceanTerritory,
-            CountryAlpha2::BN => Self::BruneiDarussalam,
-            CountryAlpha2::BG => Self::Bulgaria,
-            CountryAlpha2::BF => Self::BurkinaFaso,
-            CountryAlpha2::BI => Self::Burundi,
-            CountryAlpha2::CV => Self::CaboVerde,
-            CountryAlpha2::KH => Self::Cambodia,
-            CountryAlpha2::CM => Self::Cameroon,
-            CountryAlpha2::CA => Self::Canada,
-            CountryAlpha2::KY => Self::CaymanIslands,
-            CountryAlpha2::CF => Self::CentralAfricanRepublic,
-            CountryAlpha2::TD => Self::Chad,
-            CountryAlpha2::CL => Self::Chile,
-            CountryAlpha2::CN => Self::China,
-            CountryAlpha2::CX => Self::ChristmasIsland,
-            CountryAlpha2::CC => Self::CocosKeelingIslands,
-            CountryAlpha2::CO => Self::Colombia,
-            CountryAlpha2::KM => Self::Comoros,
-            CountryAlpha2::CG => Self::Congo,
-            CountryAlpha2::CD => Self::CongoDemocraticRepublic,
-            CountryAlpha2::CK => Self::CookIslands,
-            CountryAlpha2::CR => Self::CostaRica,
-            CountryAlpha2::CI => Self::CotedIvoire,
-            CountryAlpha2::HR => Self::Croatia,
-            CountryAlpha2::CU => Self::Cuba,
-            CountryAlpha2::CW => Self::Curacao,
-            CountryAlpha2::CY => Self::Cyprus,
-            CountryAlpha2::CZ => Self::Czechia,
-            CountryAlpha2::DK => Self::Denmark,
-            CountryAlpha2::DJ => Self::Djibouti,
-            CountryAlpha2::DM => Self::Dominica,
-            CountryAlpha2::DO => Self::DominicanRepublic,
-            CountryAlpha2::EC => Self::Ecuador,
-            CountryAlpha2::EG => Self::Egypt,
-            CountryAlpha2::SV => Self::ElSalvador,
-            CountryAlpha2::GQ => Self::EquatorialGuinea,
-            CountryAlpha2::ER => Self::Eritrea,
-            CountryAlpha2::EE => Self::Estonia,
-            CountryAlpha2::ET => Self::Ethiopia,
-            CountryAlpha2::FK => Self::FalklandIslandsMalvinas,
-            CountryAlpha2::FO => Self::FaroeIslands,
-            CountryAlpha2::FJ => Self::Fiji,
-            CountryAlpha2::FI => Self::Finland,
-            CountryAlpha2::FR => Self::France,
-            CountryAlpha2::GF => Self::FrenchGuiana,
-            CountryAlpha2::PF => Self::FrenchPolynesia,
-            CountryAlpha2::TF => Self::FrenchSouthernTerritories,
-            CountryAlpha2::GA => Self::Gabon,
-            CountryAlpha2::GM => Self::Gambia,
-            CountryAlpha2::GE => Self::Georgia,
-            CountryAlpha2::DE => Self::Germany,
-            CountryAlpha2::GH => Self::Ghana,
-            CountryAlpha2::GI => Self::Gibraltar,
-            CountryAlpha2::GR => Self::Greece,
-            CountryAlpha2::GL => Self::Greenland,
-            CountryAlpha2::GD => Self::Grenada,
-            CountryAlpha2::GP => Self::Guadeloupe,
-            CountryAlpha2::GU => Self::Guam,
-            CountryAlpha2::GT => Self::Guatemala,
-            CountryAlpha2::GG => Self::Guernsey,
-            CountryAlpha2::GN => Self::Guinea,
-            CountryAlpha2::GW => Self::GuineaBissau,
-            CountryAlpha2::GY => Self::Guyana,
-            CountryAlpha2::HT => Self::Haiti,
-            CountryAlpha2::HM => Self::HeardIslandAndMcDonaldIslands,
-            CountryAlpha2::VA => Self::HolySee,
-            CountryAlpha2::HN => Self::Honduras,
-            CountryAlpha2::HK => Self::HongKong,
-            CountryAlpha2::HU => Self::Hungary,
-            CountryAlpha2::IS => Self::Iceland,
-            CountryAlpha2::IN => Self::India,
-            CountryAlpha2::ID => Self::Indonesia,
-            CountryAlpha2::IR => Self::IranIslamicRepublic,
-            CountryAlpha2::IQ => Self::Iraq,
-            CountryAlpha2::IE => Self::Ireland,
-            CountryAlpha2::IM => Self::IsleOfMan,
-            CountryAlpha2::IL => Self::Israel,
-            CountryAlpha2::IT => Self::Italy,
-            CountryAlpha2::JM => Self::Jamaica,
-            CountryAlpha2::JP => Self::Japan,
-            CountryAlpha2::JE => Self::Jersey,
-            CountryAlpha2::JO => Self::Jordan,
-            CountryAlpha2::KZ => Self::Kazakhstan,
-            CountryAlpha2::KE => Self::Kenya,
-            CountryAlpha2::KI => Self::Kiribati,
-            CountryAlpha2::KP => Self::KoreaDemocraticPeoplesRepublic,
-            CountryAlpha2::KR => Self::KoreaRepublic,
-            CountryAlpha2::KW => Self::Kuwait,
-            CountryAlpha2::KG => Self::Kyrgyzstan,
-            CountryAlpha2::LA => Self::LaoPeoplesDemocraticRepublic,
-            CountryAlpha2::LV => Self::Latvia,
-            CountryAlpha2::LB => Self::Lebanon,
-            CountryAlpha2::LS => Self::Lesotho,
-            CountryAlpha2::LR => Self::Liberia,
-            CountryAlpha2::LY => Self::Libya,
-            CountryAlpha2::LI => Self::Liechtenstein,
-            CountryAlpha2::LT => Self::Lithuania,
-            CountryAlpha2::LU => Self::Luxembourg,
-            CountryAlpha2::MO => Self::Macao,
-            CountryAlpha2::MK => Self::MacedoniaTheFormerYugoslavRepublic,
-            CountryAlpha2::MG => Self::Madagascar,
-            CountryAlpha2::MW => Self::Malawi,
-            CountryAlpha2::MY => Self::Malaysia,
-            CountryAlpha2::MV => Self::Maldives,
-            CountryAlpha2::ML => Self::Mali,
-            CountryAlpha2::MT => Self::Malta,
-            CountryAlpha2::MH => Self::MarshallIslands,
-            CountryAlpha2::MQ => Self::Martinique,
-            CountryAlpha2::MR => Self::Mauritania,
-            CountryAlpha2::MU => Self::Mauritius,
-            CountryAlpha2::YT => Self::Mayotte,
-            CountryAlpha2::MX => Self::Mexico,
-            CountryAlpha2::FM => Self::MicronesiaFederatedStates,
-            CountryAlpha2::MD => Self::MoldovaRepublic,
-            CountryAlpha2::MC => Self::Monaco,
-            CountryAlpha2::MN => Self::Mongolia,
-            CountryAlpha2::ME => Self::Montenegro,
-            CountryAlpha2::MS => Self::Montserrat,
-            CountryAlpha2::MA => Self::Morocco,
-            CountryAlpha2::MZ => Self::Mozambique,
-            CountryAlpha2::MM => Self::Myanmar,
-            CountryAlpha2::NA => Self::Namibia,
-            CountryAlpha2::NR => Self::Nauru,
-            CountryAlpha2::NP => Self::Nepal,
-            CountryAlpha2::NL => Self::Netherlands,
-            CountryAlpha2::NC => Self::NewCaledonia,
-            CountryAlpha2::NZ => Self::NewZealand,
-            CountryAlpha2::NI => Self::Nicaragua,
-            CountryAlpha2::NE => Self::Niger,
-            CountryAlpha2::NG => Self::Nigeria,
-            CountryAlpha2::NU => Self::Niue,
-            CountryAlpha2::NF => Self::NorfolkIsland,
-            CountryAlpha2::MP => Self::NorthernMarianaIslands,
-            CountryAlpha2::NO => Self::Norway,
-            CountryAlpha2::OM => Self::Oman,
-            CountryAlpha2::PK => Self::Pakistan,
-            CountryAlpha2::PW => Self::Palau,
-            CountryAlpha2::PS => Self::PalestineState,
-            CountryAlpha2::PA => Self::Panama,
-            CountryAlpha2::PG => Self::PapuaNewGuinea,
-            CountryAlpha2::PY => Self::Paraguay,
-            CountryAlpha2::PE => Self::Peru,
-            CountryAlpha2::PH => Self::Philippines,
-            CountryAlpha2::PN => Self::Pitcairn,
-            CountryAlpha2::PL => Self::Poland,
-            CountryAlpha2::PT => Self::Portugal,
-            CountryAlpha2::PR => Self::PuertoRico,
-            CountryAlpha2::QA => Self::Qatar,
-            CountryAlpha2::RE => Self::Reunion,
-            CountryAlpha2::RO => Self::Romania,
-            CountryAlpha2::RU => Self::RussianFederation,
-            CountryAlpha2::RW => Self::Rwanda,
-            CountryAlpha2::BL => Self::SaintBarthelemy,
-            CountryAlpha2::SH => Self::SaintHelenaAscensionAndTristandaCunha,
-            CountryAlpha2::KN => Self::SaintKittsAndNevis,
-            CountryAlpha2::LC => Self::SaintLucia,
-            CountryAlpha2::MF => Self::SaintMartinFrenchpart,
-            CountryAlpha2::PM => Self::SaintPierreAndMiquelon,
-            CountryAlpha2::VC => Self::SaintVincentAndTheGrenadines,
-            CountryAlpha2::WS => Self::Samoa,
-            CountryAlpha2::SM => Self::SanMarino,
-            CountryAlpha2::ST => Self::SaoTomeAndPrincipe,
-            CountryAlpha2::SA => Self::SaudiArabia,
-            CountryAlpha2::SN => Self::Senegal,
-            CountryAlpha2::RS => Self::Serbia,
-            CountryAlpha2::SC => Self::Seychelles,
-            CountryAlpha2::SL => Self::SierraLeone,
-            CountryAlpha2::SG => Self::Singapore,
-            CountryAlpha2::SX => Self::SintMaartenDutchpart,
-            CountryAlpha2::SK => Self::Slovakia,
-            CountryAlpha2::SI => Self::Slovenia,
-            CountryAlpha2::SB => Self::SolomonIslands,
-            CountryAlpha2::SO => Self::Somalia,
-            CountryAlpha2::ZA => Self::SouthAfrica,
-            CountryAlpha2::GS => Self::SouthGeorgiaAndTheSouthSandwichIslands,
-            CountryAlpha2::SS => Self::SouthSudan,
-            CountryAlpha2::ES => Self::Spain,
-            CountryAlpha2::LK => Self::SriLanka,
-            CountryAlpha2::SD => Self::Sudan,
-            CountryAlpha2::SR => Self::Suriname,
-            CountryAlpha2::SJ => Self::SvalbardAndJanMayen,
-            CountryAlpha2::SZ => Self::Swaziland,
-            CountryAlpha2::SE => Self::Sweden,
-            CountryAlpha2::CH => Self::Switzerland,
-            CountryAlpha2::SY => Self::SyrianArabRepublic,
-            CountryAlpha2::TW => Self::TaiwanProvinceOfChina,
-            CountryAlpha2::TJ => Self::Tajikistan,
-            CountryAlpha2::TZ => Self::TanzaniaUnitedRepublic,
-            CountryAlpha2::TH => Self::Thailand,
-            CountryAlpha2::TL => Self::TimorLeste,
-            CountryAlpha2::TG => Self::Togo,
-            CountryAlpha2::TK => Self::Tokelau,
-            CountryAlpha2::TO => Self::Tonga,
-            CountryAlpha2::TT => Self::TrinidadAndTobago,
-            CountryAlpha2::TN => Self::Tunisia,
-            CountryAlpha2::TR => Self::Turkey,
-            CountryAlpha2::TM => Self::Turkmenistan,
-            CountryAlpha2::TC => Self::TurksAndCaicosIslands,
-            CountryAlpha2::TV => Self::Tuvalu,
-            CountryAlpha2::UG => Self::Uganda,
-            CountryAlpha2::UA => Self::Ukraine,
-            CountryAlpha2::AE => Self::UnitedArabEmirates,
-            CountryAlpha2::GB => Self::UnitedKingdomOfGreatBritainAndNorthernIreland,
-            CountryAlpha2::US => Self::UnitedStatesOfAmerica,
-            CountryAlpha2::UM => Self::UnitedStatesMinorOutlyingIslands,
-            CountryAlpha2::UY => Self::Uruguay,
-            CountryAlpha2::UZ => Self::Uzbekistan,
-            CountryAlpha2::VU => Self::Vanuatu,
-            CountryAlpha2::VE => Self::VenezuelaBolivarianRepublic,
-            CountryAlpha2::VN => Self::Vietnam,
-            CountryAlpha2::VG => Self::VirginIslandsBritish,
-            CountryAlpha2::VI => Self::VirginIslandsUS,
-            CountryAlpha2::WF => Self::WallisAndFutuna,
-            CountryAlpha2::EH => Self::WesternSahara,
-            CountryAlpha2::YE => Self::Yemen,
-            CountryAlpha2::ZM => Self::Zambia,
-            CountryAlpha2::ZW => Self::Zimbabwe,
-        }
-    }
-}
-
 #[derive(Clone, Copy, Eq, Hash, PartialEq, Debug, Serialize, Deserialize, strum::Display)]
 #[rustfmt::skip]
 pub enum CountryAlpha3 {
