+++ crates/types-traits/domain_types/src/payouts/payouts_types.rs
@@ -398,54 +398,13 @@ pub struct PayoutCreateRecipientRequest {
 }
 
 impl PayoutCreateRecipientRequest {
+    /// Navigate to the billing `AddressDetails`; per-field accessors live on
+    /// [`crate::payment_address::AddressDetails`] and are reused from there.
     pub fn get_optional_billing_address(&self) -> Option<&crate::payment_address::AddressDetails> {
         self.address
             .as_ref()
             .and_then(|a| a.billing_address.as_ref())
             .and_then(|b| b.address.as_ref())
-    }
-
-    pub fn get_optional_billing_first_name(&self) -> Option<Secret<String>> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.first_name.clone())
-    }
-
-    pub fn get_optional_billing_last_name(&self) -> Option<Secret<String>> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.last_name.clone())
-    }
-
-    pub fn get_optional_billing_line1(&self) -> Option<Secret<String>> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.line1.clone())
-    }
-
-    pub fn get_optional_billing_line2(&self) -> Option<Secret<String>> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.line2.clone())
-    }
-
-    pub fn get_optional_billing_city(&self) -> Option<String> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.city.as_ref())
-            .map(|c| c.peek().clone())
-    }
-
-    pub fn get_optional_billing_zip(&self) -> Option<Secret<String>> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.zip.clone())
-    }
-
-    pub fn get_optional_billing_country(&self) -> Option<common_enums::CountryAlpha2> {
-        self.get_optional_billing_address()
-            .and_then(|addr| addr.country)
-    }
-
-    pub fn get_optional_billing_email(&self) -> Option<common_utils::pii::Email> {
-        self.address
-            .as_ref()
-            .and_then(|a| a.billing_address.as_ref())
-            .and_then(|b| b.email.clone())
     }
 }
